completed

+++ routes/category.js
@@ -1,87 +1,134 @@
 const express = require('express');
 const router = express.Router();
 const Category = require('../models/Category'); // Import your Mongoose model
+const jwt = require("jsonwebtoken")
+const dotenv = require("dotenv");
+
+dotenv.config();
+
+const secretKey = process.env.JWT_SECRET_KEY
+
+
+const verify = (req, res, next) => {
+  const authHeader = req.headers.authorization;
+  if (authHeader) {
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token, secretKey, (err, user) => {
+      if (err) {
+        return res.status(401).json({ error: "Token is not valid" });
+      }
+      req.user = user;
+      next();
+    })
+  } else {
+    res.status(400).json({ error: "You are not authenticated" });
+  }
+}
 
 // Handle POST request to create a new category
-router.post('/create', async (req, res) => {
-  try {
-    // Extract data from the request body
-    let { name, sections, slug } = req.body;
+router.post('/create', verify, async (req, res) => {
+  if (req.user.role === "superadmin") {
+    try {
+      // Extract data from the request body
+      let { category_type, sections, slug } = req.body;
 
-    // Convert the name and slug to lowercase for case-insensitive comparison
-    name = name.toLowerCase();
+      // Convert the name and slug to lowercase for case-insensitive comparison
+      category_type = category_type.toLowerCase();
 
-    // Check if a category with the same name already exists
-    const existingNameCategory = await Category.findOne({ name });
-    if (existingNameCategory) {
-      return res.status(400).json({ error: 'Category with the same name already exists' });
+      // Check if a category with the same name already exists
+      const existingNameCategory = await Category.findOne({ category_type });
+      if (existingNameCategory) {
+        return res.status(400).json({ error: 'Category with the same name already exists' });
+      }
+
+      // Create a new Category document
+      const newCategory = new Category({
+        category_type,
+        slug,
+        sections,
+      });
+
+      // Save the document to the database
+      await newCategory.save();
+
+      // Respond with the saved category
+      res.status(201).json({ message: 'Category created successfully' });
+    } catch (error) {
+      console.log(error)
+      // Handle errors
+      if (error.code === 11000 && error.keyPattern && error.keyPattern.category_type) {
+        res.status(400).json({ error: 'Category already exists' })
+      } else if (error.code === 11000 && error.keyPattern && error.keyPattern.slug) {
+        res.status(400).json({ error: "Slug should be unique for each category" })
+      }
+      else {
+        res.status(500).json({ error: "Internal Server Error" })
+      }
     }
+  } else {
+    res.status(403).json({ error: "No Access to perform this action" })
+  }
 
-    // Create a new Category document
-    const newCategory = new Category({
-      name,
-      slug,
-      sections,
-    });
+});
 
-    // Save the document to the database
-    await newCategory.save();
-
-    // Respond with the saved category
-    res.status(201).json({ message: 'Category created successfully' });
+router.get("/get-all-category-types", async (req, res) => {
+  try {
+    // Fetch all categories from the database and select only the category_type field
+    const allCategoryTypes = await Category.find().select('_id category_type');
+    // Extract the category_type values from the array of documents
+    // Respond with the array of category_type values
+    res.status(200).json(allCategoryTypes);
   } catch (error) {
     // Handle errors
-    if (error.code === 11000 && error.keyPattern && error.keyPattern.name) {
-      res.status(400).json({ error: 'Category already exists' })
-    } else if (error.code === 11000 && error.keyPattern && error.keyPattern.slug) {
-      res.status(400).json({ error: "Slug should be unique for each category" })
-    }
-    else {
-      res.status(500).json({ error: "Internal Server Error" })
-    }
+    res.status(500).json({ error: 'Internal Server Error' });
   }
 });
 
 // Handle PUT request to edit a category by ID
-router.put('/edit-category/:categoryId', async (req, res) => {
-  try {
-    const categoryId = req.params.categoryId;
-    const { name, sections, slug } = req.body;
+router.put('/edit-category/:categoryId', verify, async (req, res) => {
+  if (req.user.role === "superadmin") {
+    try {
+      const categoryId = req.params.categoryId;
+      const { category_type, sections, slug } = req.body;
 
-    // Convert the name and slug to lowercase for case-insensitive comparison
-    const lowercasedName = name.toLowerCase();
+      // Convert the name and slug to lowercase for case-insensitive comparison
+      const lowercasedName = category_type.toLowerCase();
 
-    // Check if a category with the same name already exists (excluding the current category being edited)
-    const existingNameCategory = await Category.findOne({ name: lowercasedName, _id: { $ne: categoryId } });
-    if (existingNameCategory) {
-      return res.status(400).json({ error: 'Category with the same name already exists' });
+      // Check if a category with the same name already exists (excluding the current category being edited)
+      const existingNameCategory = await Category.findOne({ category_type: lowercasedName, _id: { $ne: categoryId } });
+      if (existingNameCategory) {
+        return res.status(400).json({ error: 'Category with the same name already exists' });
+      }
+
+      // Find the category by ID and update its fields
+      const updatedCategory = await Category.findByIdAndUpdate(categoryId, {
+        category_type: lowercasedName,
+        slug,
+        sections,
+      }, { new: true });
+
+      // Check if the category with the given ID exists
+      if (!updatedCategory) {
+        return res.status(404).json({ error: 'Category not found' });
+      }
+
+      // Respond with the updated category
+      res.json({ message: 'Category updated successfully', category: updatedCategory });
+    } catch (error) {
+      console.log(error)
+      // Handle errors
+      if (error.code === 11000 && error.keyPattern && error.keyPattern.category_type) {
+        res.status(400).json({ error: 'Category already exists' });
+      } else if (error.code === 11000 && error.keyPattern && error.keyPattern.slug) {
+        res.status(400).json({ error: 'Slug should be unique for each category' });
+      } else {
+        res.status(500).json({ error: 'Internal Server Error' });
+      }
     }
-
-    // Find the category by ID and update its fields
-    const updatedCategory = await Category.findByIdAndUpdate(categoryId, {
-      name: lowercasedName,
-      slug,
-      sections,
-    }, { new: true });
-
-    // Check if the category with the given ID exists
-    if (!updatedCategory) {
-      return res.status(404).json({ error: 'Category not found' });
-    }
-
-    // Respond with the updated category
-    res.json({ message: 'Category updated successfully', category: updatedCategory });
-  } catch (error) {
-    console.log(error)
-    // Handle errors
-    if (error.code === 11000 && error.keyPattern && error.keyPattern.name) {
-      res.status(400).json({ error: 'Category already exists' });
-    } else if (error.code === 11000 && error.keyPattern && error.keyPattern.slug) {
-      res.status(400).json({ error: 'Slug should be unique for each category' });
-    } else {
-      res.status(500).json({ error: 'Internal Server Error' });
-    }
+  } else {
+    res.status(403).json({ error: "No Access to perform this action" })
   }
+
 });
 
 
@@ -97,23 +144,28 @@ router.get('/fetch-all-categories', async (req, res) => {
 });
 
 
-router.delete('/delete/:categoryId', async (req, res) => {
-  try {
-    const { categoryId } = req.params;
+router.delete('/delete/:categoryId', verify, async (req, res) => {
+  if (req.user.role === "superadmin") {
+    try {
+      const { categoryId } = req.params;
 
-    // Attempt to find and delete the category by ID
-    const deletedCategory = await Category.findByIdAndDelete(categoryId);
+      // Attempt to find and delete the category by ID
+      const deletedCategory = await Category.findByIdAndDelete(categoryId);
 
-    // Check if the category was found and deleted
-    if (!deletedCategory) {
-      return res.status(404).json({ error: 'Category not found' });
+      // Check if the category was found and deleted
+      if (!deletedCategory) {
+        return res.status(404).json({ error: 'Category not found' });
+      }
+
+      res.json({ message: 'Category deleted successfully' });
+    } catch (error) {
+      console.error(error);
+      res.status(500).json({ error: 'Internal Server Error' });
     }
-
-    res.json({ message: 'Category deleted successfully' });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Internal Server Error' });
+  } else {
+    res.status(403).json({ error: "No Access to perform this action" })
   }
+
 });
 
 router.get('/fetch-category-id/:categoryId', async (req, res) => {
@@ -141,7 +193,7 @@ router.get('/fetch-category-name/:categoryName', async (req, res) => {
     const categoryName = req.params.categoryName;
 
     // Find the category by name
-    const category = await Category.findOne({ name: categoryName });
+    const category = await Category.findOne({ category_type: categoryName });
 
     // Check if the category with the given name exists
     if (!category) {
@@ -171,7 +223,7 @@ router.get('/get-all-categories', async (req, res) => {
     }
 
     const allCategories = await Category.find(query)
-      .select('name slug sections') // Include the 'sections' field in the selection
+      .select('category_type slug sections') // Include the 'sections' field in the selection
       .sort({ createdAt: -1 })
       .skip(skip)
       .limit(parseInt(pageSize));
@@ -181,7 +233,7 @@ router.get('/get-all-categories', async (req, res) => {
     // Create a new array with only the count of sections for each category
     const sectionCounts = allCategories.map(category => ({
       id: category._id,
-      name: category.name,
+      category_type: category.category_type,
       slug: category.slug,
       sectionCount: category.sections ? category.sections.length : 0,
     }));
